+++ src/App.jsx
@@ -1,11 +1,25 @@
-import { Route, Routes } from "react-router-dom";
+import { Route, Routes, useSearchParams } from "react-router-dom";
 import NotFoundPage from "./pages/NotFoundPage";
 import HomePage from "./pages/HomePage";
 import LoginPage from "./pages/LoginPage";
 import RegisterPage from "./pages/RegisterPage";
 import { URL_CONSTANTS } from "./constants/url.constants";
+import { useEffect } from "react";
 
 function App() {
+  const [params] = useSearchParams();
+  useEffect(() => {
+    const accessToken = params.get("accessToken");
+    const refreshToken = params.get("refreshToken");
+    const newUser = params.get("newUser");
+    if (newUser === "false") {
+      alert("Login");
+    } else {
+      alert("Đăng Ký");
+    }
+    localStorage.setItem("access_token", accessToken);
+    localStorage.setItem("refresh_token", refreshToken);
+  }, [params]);
   return (
     <Routes>
       {/* {user && user.role === "ADMIN" ? ( */}
